changing np.int32-->np.int64 in many cases to stop negatives appearing

+++ epilogos/roiAndVisualPairwise.py
@@ -334,7 +334,7 @@ def readInData(outputDirPath, numProcesses, numStates):
     chrDict    = dict(zip(chrNumbers, chrOrder))
 
     # Convert dataframes to np arrays for easier manipulation
-    locationArr = diffDF.iloc[:, 0:3].replace({"chr": dict(zip(chrOrder, chrNumbers))}).to_numpy(dtype=np.int32)
+    locationArr = diffDF.iloc[:, 0:3].replace({"chr": dict(zip(chrOrder, chrNumbers))}).to_numpy(dtype=np.int64)
     diffArr     = diffDF.iloc[:, 3:].to_numpy(dtype=np.float32)
 
     # Cleaning up the temp files after we've read them
